+++ adapter/src/adapter/api/Endpoint.dto.ts
