+++ adapter/src/adapter/model/AdapterConfig.ts
@@ -1,6 +1,7 @@
 import { validators } from '@jvalue/node-dry-basics';
 
-import { AdapterConfigDTO } from './Endpoint.dto';
+import { AdapterConfigDTO } from '../api/Endpoint.dto';
+
 import { FormatConfig } from './FormatConfig';
 import { ProtocolConfig } from './ProtocolConfig';
