+++ adapter/src/datasource/services/dataImportTriggerService.ts
@@ -7,10 +7,10 @@ import { ProtocolConfig } from '../../adapter/model/ProtocolConfig';
 import { Protocol } from '../../adapter/Protocol';
 import { AdapterService } from '../../adapter/services/adapterService';
 import { ADAPTER_AMQP_IMPORT_SUCCESS_TOPIC } from '../../env';
-import { DataImportDTO, dataimportEntityToDTO } from '../model/DataImport.dto';
+import { DataImportDTO, dataimportEntityToDTO } from '../api/DataImport.dto';
+import { DatasourceDTO, datasourceEntityToDTO } from '../api/Datasource.dto';
 import { DataImportEntity } from '../model/DataImport.entity';
 import { DataImportInsertStatement } from '../model/DataImportInsertStatement';
-import { DatasourceDTO, datasourceEntityToDTO } from '../model/Datasource.dto';
 import { DatasourceEntity } from '../model/Datasource.entity';
 import { DataImportRepository } from '../repository/dataImportRepository';
 import { DatasourceRepository } from '../repository/datasourceRepository';
